Split Program.cs into Program.cs and TflApi.cs Extracted PromptUserAndGetInputFromConsole from Main
