+++ BusBoard.ConsoleApp/BusBoard.ConsoleApp.csproj
@@ -63,6 +63,7 @@
     <Compile Include="PostcodeApiResponse.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TflApi.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
