+++ BusBoard.ConsoleApp/Program.cs
@@ -13,23 +13,30 @@ namespace BusBoard.ConsoleApp
   {
     static void Main(string[] args)
     {
+      TflApi api = new TflApi();
+      PromptUserAndGetInputFromConsole(api);
+    }
+
+    private static void PromptUserAndGetInputFromConsole(TflApi api)
+    {
       Console.WriteLine("Postcode or Stop point ID? (P or S)");
-      switch (Console.ReadLine())
+      switch (Console.ReadLine().ToUpper())
       {
         case "S":
           Console.WriteLine("Please enter your desired stop point ID:");
           var stopCode = Console.ReadLine();
           Console.WriteLine(" ");
-          PrintBusesFromStopCode(stopCode, "");
+          PrintBusesFromStopCode(stopCode, "", api);
           break;
         case "P":
           Console.WriteLine("Please enter your desired postcode");
           var postcode = Console.ReadLine();
           Console.WriteLine(" ");
-          foreach (var sc in GetTwoClosestBusStopsToPostcode(postcode))
+          foreach (var sc in api.GetTwoClosestBusStopsToPostcode(postcode))
           {
-            PrintBusesFromStopCode(sc.NaptanId, sc.CommonName);
+            PrintBusesFromStopCode(sc.NaptanId, sc.CommonName, api);
           }
+
           break;
         default:
           Console.WriteLine("Please enter S or P");
@@ -37,33 +44,9 @@ namespace BusBoard.ConsoleApp
       }
     }
 
-    private static IEnumerable<BusStop> GetTwoClosestBusStopsToPostcode(string postcodeString)
+    private static void PrintBusesFromStopCode(string stopCode, string commonName, TflApi api)
     {
-      var requestUrl = "http://api.postcodes.io";
-      var client = new RestClient(requestUrl);
-      var request = new RestRequest($"postcodes/{postcodeString}", DataFormat.Json);
-      var response = client.Execute(request);
-      var postcodeApiResponse = JsonConvert.DeserializeObject<PostcodeApiResponse>(response.Content);
-      var postcode = postcodeApiResponse.Result;
-      List<BusStop> nearbyStops = GetBusStopsNearPostcode(postcode);
-      IEnumerable<BusStop> closestTwoStops = nearbyStops.OrderBy(stop => stop.Distance).Take(2);
-      return closestTwoStops;
-    }
-
-    private static List<BusStop> GetBusStopsNearPostcode(PostcodeLocation postcodeLocation)
-    {
-      var requestUrl = "https://api.tfl.gov.uk";
-      var client = new RestClient(requestUrl);
-      var request = new RestRequest($"StopPoint?stopTypes=NaptanPublicBusCoachTram&radius=200&useStopPointHierarchy=false&returnLines=false&lat={postcodeLocation.Latitude}&lon={postcodeLocation.Longitude}", DataFormat.Json);
-      var response = client.Execute(request);
-      var busStopApiResponse = JsonConvert.DeserializeObject<BusStopApiResponse>(response.Content);
-      return busStopApiResponse.StopPoints;
-    }
-    
-    private static void PrintBusesFromStopCode(string stopCode, string commonName)
-    {
-      var response = GetListOfArrivalPredictionsForStopPoint(stopCode);
-      var buses = GetNext5BusesFromApiResponse(response);
+      var buses = api.GetListOfArrivalPredictionsForStopPoint(stopCode);
       Console.WriteLine($"Next 5 Buses at stop {stopCode} - {commonName}");
       foreach (var bus in buses)
       {
@@ -74,16 +57,7 @@ namespace BusBoard.ConsoleApp
       Console.WriteLine(" ");
     }
     
-
-    private static IEnumerable<BusArrivalPrediction> GetNext5BusesForStopPoint(string stopCode)
-    {
-      var requestUrl = "https://api.tfl.gov.uk";
-      var client = new RestClient(requestUrl);
-      var request = new RestRequest($"StopPoint/{stopCode}/Arrivals");
-      var buses = client.Execute<List<BusArrivalPrediction>>(request).Data;
-      IEnumerable<BusArrivalPrediction> orderedBuses = buses.OrderBy(bus => bus.ExpectedArrival).Take(5);
-      return orderedBuses;
-    }
+    
     
   }
 }
